EES-4376 Move getMethodologyApprovalStatusLabel to methodologyUtil.ts
